Switches to 32 bits per sample for https://github.com/atomic14/esp32-tft-audio-monitor/issues/2

+++ src/I2S/I2SSampler.h
@@ -35,7 +35,7 @@ private:
     i2s_pin_config_t m_i2sPins;
 
 protected:
-    void addSample(int16_t sample);
+    void addSample(int32_t sample);
 
 public:
     int16_t getBufferSizeInBytes()
